LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/fill-and-stroke-styles/2d.gradient.radial.touch2.worker-expected.txt


FAIL 2d assert_equals: Green channel of the pixel at (98, 25) expected 255 but got 254
